Move master detector into its own package.
