+++ settings.gradle
@@ -3,6 +3,7 @@ rootProject.name = 'usi'
 include 'core'
 include 'core-models'
 include 'mesos-client'
+include 'mesos-master-detector'
 include 'metrics'
 include 'metrics-dropwizard'
 include 'persistence'
